suggested updates in PR request...move stats update out of print_stats
Former-commit-id: 5d0ad924aacd0967cddaf8107b31a20f0a35791f [formerly 474e2f13512812ff5a366fbfe826ff2cad834c54] [formerly 411635256a86e85c0eedf3aa9f0408a5399f23e1 [formerly 22c21b22688a313610baa6ffdcdee896ad237516]]
Former-commit-id: ee8aec779a24ba111acb74f6e53db23427f606e1 [formerly 2902480fdbd2411d25952318f9726de19a9dd5a0]
Former-commit-id: 0881f8627a99119a7d1a4e882f1974bd8f2c1304

+++ src/util/utils.cc
@@ -341,10 +341,10 @@ namespace FT{
                                float bst_score_v,
                                float md_score,
                                float md_loss_v,
-                               float md_size,
-                               float md_complexity,
-                               float md_num_params,
-                               float md_dim)
+                               unsigned md_size,
+                               unsigned md_complexity,
+                               unsigned md_num_params,
+                               unsigned md_dim)
         {
             generation.push_back(index+1);
             time.push_back(timer_count);
